+++ examples/jit/gcn.py
@@ -18,8 +18,8 @@ class GCN(torch.nn.Module):
     def __init__(self, in_channels: int, hidden_channels: int,
                  out_channels: int):
         super().__init__()
-        self.conv1 = GCNConv(in_channels, hidden_channels).jittable()
-        self.conv2 = GCNConv(hidden_channels, out_channels).jittable()
+        self.conv1 = GCNConv(in_channels, hidden_channels)
+        self.conv2 = GCNConv(hidden_channels, out_channels)
 
     def forward(self, x: Tensor, edge_index: Tensor) -> Tensor:
         x = F.dropout(x, p=0.5, training=self.training)