+++ test/nn/conv/test_graph_conv.py
@@ -37,7 +37,7 @@ def test_graph_conv():
         assert torch.allclose(conv(x1, adj4.t()), out2, atol=1e-6)
 
     if is_full_test():
-        jit = torch.jit.script(conv.jittable())
+        jit = torch.jit.script(conv)
         assert torch.allclose(jit(x1, edge_index), out1)
         assert torch.allclose(jit(x1, edge_index, size=(4, 4)), out1)
         assert torch.allclose(jit(x1, edge_index, value), out2)
@@ -80,7 +80,7 @@ def test_graph_conv():
         assert torch.allclose(conv((x1, None), adj4.t()), out4, atol=1e-6)
 
     if is_full_test():
-        jit = torch.jit.script(conv.jittable())
+        jit = torch.jit.script(conv)
         assert torch.allclose(jit((x1, x2), edge_index), out1)
         assert torch.allclose(jit((x1, x2), edge_index, size=(4, 2)), out1)
         assert torch.allclose(jit((x1, None), edge_index, size=(4, 2)), out2)