+++ test/nn/models/test_linkx.py
@@ -25,7 +25,7 @@ def test_linkx(num_edge_layers):
         assert torch.allclose(out, model(x, adj.t()), atol=1e-6)
 
     if is_full_test():
-        jit = torch.jit.script(model.jittable())
+        jit = torch.jit.script(model)
         assert torch.allclose(jit(x, edge_index), out)
 
         if torch_geometric.typing.WITH_TORCH_SPARSE: